refactor: use named routes

+++ routes/web.php
@@ -36,9 +36,9 @@ Route::get('ping', function () {
 });
 
 Route::get('/', [PostController::class, 'index'])->name('home');
-Route::get('posts/{post:slug}', [PostController::class, 'show']);
+Route::get('posts/{post:slug}', [PostController::class, 'show'])->name('show');
 
-Route::post('posts/{post:slug}/comments', [PostCommentController::class, 'store']);
+Route::post('posts/{post:slug}/comments', [PostCommentController::class, 'store'])->name('posts.store');
 
 Route::get('register', [RegisterController::class, 'create'])->middleware('guest');
 Route::post('register', [RegisterController::class, 'store'])->middleware('guest');
@@ -48,9 +48,9 @@ Route::post('login', [SessionsController::class, 'store'])->middleware('guest');
 Route::post('logout', [SessionsController::class, 'destroy'])->middleware('auth');
 
 // Admin
-Route::post('admin/posts', [AdminPostController::class, 'store'])->middleware('admin');
-Route::get('admin/posts/create', [AdminPostController::class, 'create'])->middleware('admin');
-Route::get('admin/posts', [AdminPostController::class, 'index'])->middleware('admin');
-Route::get('admin/posts/{post}/edit', [AdminPostController::class, 'edit'])->middleware('admin');
-Route::patch('admin/posts/{post}', [AdminPostController::class, 'update'])->middleware('admin');
-Route::delete('admin/posts/{post}', [AdminPostController::class, 'destroy'])->middleware('admin');
+Route::post('admin/posts', [AdminPostController::class, 'store'])->middleware('admin')->name('admin.posts.store');
+Route::get('admin/posts/create', [AdminPostController::class, 'create'])->middleware('admin')->name('admin.posts.create');
+Route::get('admin/posts', [AdminPostController::class, 'index'])->middleware('admin')->name('admin.posts');
+Route::get('admin/posts/{post}/edit', [AdminPostController::class, 'edit'])->middleware('admin')->name('admin.posts.edit');
+Route::patch('admin/posts/{post}', [AdminPostController::class, 'update'])->middleware('admin')->name('admin.posts.update');
+Route::delete('admin/posts/{post}', [AdminPostController::class, 'destroy'])->middleware('admin')->name('admin.posts.destroy');
